phase1252-clang-tidy-WAE-sweep-14rule-set-to-0(149-fixed+9-NOLINT-sealed)+broader-debt-ADR-MARATHON-FINAL
ALL-MODULES-TO-100 marathon FINAL item (bands 0-7 = phase1228-1251 brought
all 122 engine libs to terminal-100%; this closes the queued lint sweep).

14-rule test-suite WAE debt (measured 245 diag / 103 files) -> 0:
  * 149 behaviour-preserving fixes: parallel autofix on 39 test .cpp
    (header-filter=none, golden-safe) + 7 group-partitioned code-consistency
    agents hand-editing the rest (use-ranges/use-auto/isolate-decl/
    implicit-widening/unused-using/using-namespace->using-decl/emplace/
    namespace-comment/raw-string/prefer-member-init).
  * 9 NOLINTNEXTLINE seals (no clean idiomatic form): 8 heterogeneous-comparator
    std::lower_bound (CameraPath/CurveTrack/EventTrack/SelectionSet) + 1 braced
    std::pair emplace (Surround.hpp).

Broader 130 WAE across 24 OUT-OF-SCOPE rules (misplaced-widening-cast,
duplicate-include, use-scoped-lock, move-const-arg, member-init, ...) + 621
non-WAE advisory warnings (identifier-naming, unchecked-optional-access) SEALED
promote-on-need in docs/ADR/ADR-20260616-clang-tidy-debt-sweep.md.

Gate: cmake --build --preset ninja-debug clean -Werror (0 err / 0 warn);
ctest --preset ninja-debug -E rhi_vulkan = 317/317 PASS;
golden/sponza/chrome byte-identical (re-verified post-seal). No render regression.

Breakage fixes during sweep: texture_compress test CompressedTexture using-decl
(using-namespace->decl conversion missed a symbol); dsp_fx test std::iota kept
classic (std::ranges::iota requires a weakly_incrementable seed, float is not).

+++ engine/asset/tests/image/test_asset_image.cpp
@@ -267,7 +267,7 @@ TEST(AssetImage, GenerateMipsFullChainOn256Square)
     src.width = 256;
     src.height = 256;
     src.has_alpha = true;
-    src.rgba.assign(256U * 256U * 4U, 128U);  // solid gray
+    src.rgba.assign(static_cast<std::size_t>(256U) * 256U * 4U, 128U);  // solid gray
     auto r = cd::asset::image::generate_mips(src);
     ASSERT_TRUE(r.has_value());
     // 256 → 128 → 64 → 32 → 16 → 8 → 4 → 2 → 1 = 9 levels.
@@ -285,7 +285,7 @@ TEST(AssetImage, GenerateMipsCappedLevels)
     cd::asset::image::Image src;
     src.width = 32;
     src.height = 32;
-    src.rgba.assign(32U * 32U * 4U, 200U);
+    src.rgba.assign(static_cast<std::size_t>(32U) * 32U * 4U, 200U);
     auto r = cd::asset::image::generate_mips(src, 3);
     ASSERT_TRUE(r.has_value());
     EXPECT_EQ(r->size(), 3U);
@@ -300,7 +300,7 @@ TEST(AssetImage, GenerateMipsNonSquareWidthOrHeightOf1)
     cd::asset::image::Image src;
     src.width = 4;
     src.height = 1;
-    src.rgba.assign(4U * 1U * 4U, 99U);
+    src.rgba.assign(static_cast<std::size_t>(4U) * 1U * 4U, 99U);
     auto r = cd::asset::image::generate_mips(src);
     ASSERT_TRUE(r.has_value());
     EXPECT_EQ(r->size(), 3U);
